refactor: move ExampleResponses to features/ folder
