--- packages/api-reference/src/components/Content/Operation/PathResponses/index.ts
@@ -1,1 +1,0 @@
-export { default as PathResponses } from './PathResponses.vue'
