+++ packages/api-reference/src/features/ExampleResponses/ExampleResponses.vue
@@ -4,18 +4,17 @@ import { normalizeMimeTypeObject } from '@scalar/oas-utils/helpers'
 import type { TransformedOperation } from '@scalar/types/legacy'
 import { computed, ref } from 'vue'
 
-import { ExamplePicker } from '../../../../features/ExampleRequest'
-import { useClipboard } from '../../../../hooks'
 import {
   Card,
   CardContent,
   CardFooter,
   CardTab,
   CardTabHeader,
-} from '../../../Card'
+} from '../../components/Card'
+import { useClipboard } from '../../hooks'
+import { ExamplePicker } from '../ExampleRequest'
 import ExampleResponse from './ExampleResponse.vue'
-
-// import Headers from './Headers.vue'
+import Headers from './Headers.vue'
 
 /**
  * TODO: copyToClipboard isn’t using the right content if there are multiple examples
@@ -127,12 +126,11 @@ const showSchema = ref(false)
       </template>
     </CardTabHeader>
     <div class="scalar-card-container custom-scroll">
-      <!-- Commenting out until we re-organize cause of height issues -->
-      <!-- <CardContent
-        v-if="currentResponse.headers"
+      <CardContent
+        v-if="currentResponse?.headers"
         muted>
         <Headers :headers="currentResponse.headers" />
-      </CardContent> -->
+      </CardContent>
       <CardContent muted>
         <template v-if="currentJsonResponse?.schema">
           <ScalarCodeBlock
