+++ packages/api-reference/src/features/ExampleResponses/index.ts
@@ -1,0 +1,1 @@
+export { default as ExampleResponses } from './ExampleResponses.vue'
